+++ packages/android_intent/example/test_driver/android_intent_e2e.dart
@@ -50,12 +50,12 @@ void main() {
     }));
   }, skip: !Platform.isAndroid);
 
-  testWidgets('#canResolve throws when no Activity is found',
+  testWidgets('#canResolveActivity throws when no Activity is found',
       (WidgetTester tester) async {
     // We can't test that any of this is really working, this is mostly just
     // checking that the plugin API is registered. Only works on Android.
     const AndroidIntent intent =
         AndroidIntent(action: 'LAUNCH', package: 'foobar');
-    await expectLater(() async => await intent.canResolve(), isFalse);
+    await expectLater(() async => await intent.canResolveActivity(), isFalse);
   }, skip: !Platform.isAndroid);
 }